backend/requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.24
sqlalchemy>=2.0,<2.1
python-jose[cryptography]>=3.3
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart>=0.0.6
email-validator>=2.1
pydantic-settings>=2.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
httpx>=0.24.0
pytest-cov>=4.0
prometheus-fastapi-instrumentator>=5.0
